requirements.txt
boto3==1.40.35
httpx==0.28.1
mcp==1.23.1
openai==2.9.0
protobuf==6.33.1
python-dotenv==1.2.1
together==1.5.31
google-genai>=1.0.0